Move TOGGLE_ORDER action handler to server reducer
